requirements.txt
numpy~=2.2.5
PyQt6~=6.8.1
scipy~=1.15.2
PyYAML~=6.0.2
optuna~=4.1.0
joblib~=1.4.2
pandas~=2.3.0
wandb~=0.21.4
plotly~=5.24.1
astropy~=7.0.1
lenspack~=1.0.0
matplotlib~=3.10.0
scikit-learn~=1.6.1
SciencePlots~=2.1.1

--find-links https://download.pytorch.org/whl/cu121; sys_platform == "Windows"
torch~=2.7.0
torchvision~=0.22.0

sciplots @ git+https://github.com/EthanTreg/Science-Plots@v1.0.11
netloader @ git+https://github.com/EthanTreg/PyTorch-Network-Loader@8e8e951
